refactor: add psql database in step
